Added demonstration script
Necessary to serve the demonstration application for local interaction and experimentation.

This combines the contents of the `demo` folder with the JavaScript implementation in the `lib` folder and puts everything in the `tmp/demo` folder for serving locally using Ruby's native `httpd` server. The application is launched in your default browser for interaction. The default port is 3030 but can be given a custom port using the first positional argument.

Milestone: minor